Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

I'm not sure if that flag is used in any other packages and if they are, they need to be defined there as well. But I know this is in CM9 and respectively in JB. I'm calling it good.

Patch-set: 1