Update patch set 9
Patch Set 9: Fails; Do not submit

Not sure what's going on here but this is pretty crazy for the simple addition of a system property to control a boolean value. There are Git conflict markers in the uploaded file and a bunch of weird boolean logic. Apparently it's untested, and doesn't actually resolve any issue?

Whats up?

Patch-set: 9