Update patch set 9
Patch Set 9: No score

I'm gonna fix this in a few min and leave... 
I'm not gonna embarrass myself and take all of your time reviewing my crappy changes. I'll just keep them to myself. Win win for everyone

As for this not building, I know exactly whats wrong. That line I added yesterday is using the Boolean useMeta but that bool is only declared for the method which the line from yesterday doesn't see. So I need to move bool useMeta out of the method

Patch-set: 9