Update patch set 12
Patch Set 12:

@Ricardo: So just because he's been around longer, does that make his better than mine? Just pondering the meaning of that comment
Also, there are a lot of Qualcomm patches in this code. Just because I commited it that doesn't make it mine. And what do you expect me to do? Label the author as Qualcomm?

Anyway, this confirms my feelings. Good bye coding. My code is hated and I will not continue to try at it if it is hated so...

Patch-set: 12