Update patch set 12
Patch Set 12:

Actually, yes. When we use BSP vendor code or any other code for that matter, we don't change the attributions. If it's done with a manual merge, we commit it with "-c commitid" or with "--author 'Blah <blah@blah.com>'". If you can't do this (maybe due to squashing or other reasons), the original source should at least be mentioned in the commit message.

It's very important in the open-source world to preserve authorship and attributions of code that you merge from outside sources. Code is like your resume in the software world. Just because Qualcomm is a big evil corporation doesn't change that.

Patch-set: 12