Update patch set 1
Patch Set 1:

Ah I was wondering how you did it in one line...
Didn't think it would be the same as Java, but Thanks!!
I'll change that tomorrow morning

And it's DICS_AUDIO_BLOCB in the BoardConfigs that I've been using and it's been working fine ;). But I'll investigate tomorrow.

Patch-set: 1