Update patch set 2
Patch Set 2:

I haven't actually tried to compile it, but seeing how you are ifdeffing out a constructor that's used in other places, I can't see how it can work.

Seriously please stop uploading stuff that doesn't compile.

The DirectTrack code goes pretty deep into the framework. It would be smarter to try and adapt the source code of the audio driver to your device than to get rid of it.

Patch-set: 2