Update patch set 16
Patch Set 16: Abandoned

Hi Rashed-

I know you are excited to get your device up and running with CM, but I recommend that you fork the repositories that you need to make changes to and get the device in full, working order first. Once that is done, you should contact devrel and we can review it and make recommendations. It will save both of us time. Using CM's Gerrit instance to learn how to code is probably not the best thing to do because a patchset like this makes us very skeptical about accepting your code. I think CM is a great project to work on, but I would recommend getting the basics down first.

Patch-set: 16