Update patch set 1
Patch Set 1:

I'm not sure if this is the right place to put this, but this did not fix the Audio playback error that started with the 9-26 Nightly. If you stop/pause a track for more than 4-5 seconds and try to play it again, it will show "Music Playback Error. Couldn't play the track you requested." A playlist will play through fine if you skip to the next song and never pause it, but as soon as you do (for 5+ secs), it will show the error.

Patch-set: 1