Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

This patch series (plus the changes to qcom/audio) fix the issues with LPA crashing, failing to play tracks, and not releasing wakelocks. Partially due to a bad merge from CAF because of branch inconsistencies, and also this bug.

Patch-set: 1