Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve

Adam, you are not right! JB and ICS are feeling self good on armv6 devices (at least with middle-end devices). Also many devices are working on it yet. So, if it needed for you - it very helpful for others

Patch-set: 1