Update patch set 1
Patch Set 1: No score

If it was accepted before and breaking the CTS is the criteria to not accept a patch then OK, but I'd also argue why should  a patch be accepted specifically for devices that are known to require CTS breakage elsewhere.

I'd love to have CM10.1 on my Optimus One but each new version requires more sacrifices, there has to be a cut-off point somewhere... But I'll leave my argument at that.

Patch-set: 1