Update patch set 1
Patch Set 1: No score

"Media Transfer Protocol" is intended for transferring "Media". The ".nomedia" flag is intended to identify that path as not containing media, i.e., does not contain anything that should be transferred by MTP.

Perhaps an alternative solution would be more to your liking. An alternative to ".nomedia", call it ".noscan", and respect that to prevent scanning of unnecessary paths.

The problem with the media scan as it is, is that it runs every single time the device is rebooted, and depending on the number of files, it can take many minutes to complete and cause a significant battery drain and reduction in performance in the interim. This is particularly a problem if you have CWM hash/blob backups on the device.

Patch-set: 1