Update patch set 1
Patch Set 1:

So how about this; if you aren't willing to consider anything that makes sense, let somebody else. As it stands, this function is broken. Period. I've suggested an alternative implementation that should alleviate your worries that nomedia shouldn't *actually* mean "NO MEDIA", yet not a single comment regarding that.

What constitutes "Excessive" is clearly a matter of opinion. Having uSD cards of 64 GB, would it be excessive to exceed a few hundred files? Because if you exceed 50% (the point where it makes sense to jump from the next smaller option, being 32 GB) on that, you probably have 10's or hundreds of thousands of files on it, and the better part of an hour and a battery wasted scanning it.

Patch-set: 1