Update patch set 2
Patch Set 2: Code-Review+1

@Ricardo:
The setting is actually needed and used in cm10: 
debug.camcorder.disablemeta 1 is set in 
https://github.com/CyanogenMod/andro...ean/msm7x30.mk

The /system/build.prop in a plain fxp214 install has the disablemeta setting.
It is also needed in cm10.1 for all Xperia 2011 devices.

Reason is that the Sony camera drivers for the entire 2011 line are older than dirt, and AFAIK don't support metadata.

10.1 puts the fun back into the 2011 Sonys, thanks to Mike and others. Please approve.

Patch-set: 2