Update patch set 5
Patch Set 5:

If you use Apollo oder Play Music, play an album. The first song is played via tunnel which can be seen in logcat if you seek a bit. If you let this song play to an end the next song will be played with normal audioplayer which you can also see if you seek a bit, there is different output.

Also you can set the DSP Manager to something fancy, then you can hear that DSP manager works for every second track (because it is not working for tunnelplayer at the moment)

Patch-set: 5