Update patch set 3
Patch Set 3:

yes but i think what hes asking is your input, if its something that can be changed on kernel side to work with current code or if its just a msm8660 htc thing and only feasible way to allow them to work would be this ifdef approach. I know they are old as hell. I have a jf now. but I have so many old htc devices its a shame to seem them go to waste

Patch-set: 3