Update patch set 2
Patch Set 2:

I created a wrapper for audio.primary and audio_policy blobs when I tried to get the HTC Tegra 3 4.1/4.2 audio blobs running. You can find the code here: https://github.com/thoemy/android-audio-wrapper

In theory it should be possible to wrap 4.0/4.1/4.2 blobs to be used for 4.3. Granted the necessary headers are provided and some code must be adjusted. I wanted to make it configurable though Makefiles. But I think I failed a bit in that regard and actually made it more complicated :D.

I haven't touched any 4.4 related stuff so I don't know what changed and if it it's possible to use the wrapper.
Anyway. Please give it a try and don't hesitate to contact me if you think it might be useful. I have abandoned it a little since I managed to get the 4.2 Tegra 3 blobs working.

It should be noted that you should use the stock audio_policy if possible.

Patch-set: 2