Update patch set 2
Patch Set 2:

Etan - didn't we go through this last night in relation to moving 8660 (which is HIGHLY modified) to legacy?  I think that we need to be consistent in the approach of what is and isn't supported within CM.   If we upload 8660 compatibility patches for CM11 and they keep getting shot down or "sat on" .. what's the point of having 8660 at all in CM11?  (Which brings us to the conversation we all had on IRC last night)

Patch-set: 2