Update patch set 2
Patch Set 2:

OMAP devices aren't a pain Steve, you pushing through changes without a proper review is the pain. This broke more than just OMAP as well. Your code isn't above and beyond review, and until you figure that out, mishaps like this are going to keep happening. It just seems nobody has the balls to call you out on it. If devices that break from a change are somewhat obscure I could understand, but the Galaxy Nexus is extremely popular and if you combine the total installs of maguro and toro on CM's stats it is bested only by i9100, i9300, and galaxysmtd. That's a lot of users, not to mention it's more likely for a Nexus device to be using the latest CM (11), and again, this broke more than just OMAP.

Patch-set: 2
Reviewer: Kyle Repinski <6998@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0