Update patch set 1
Patch Set 1:

So lack of review equates to apathy towards bugs and a complete shift in priorities? Is that the seriously the position you want to stake your argument on?

A simple bump or "hey where are we on this" would have sufficed. But instead you have decided to throw the equivalent of a tantrum on gerrit - even tagging in reviewers who couldn't merge this if they wanted to. 

This patch is for 10.2, which is mostly stagnant. Its stands to reason review time would not be focused there (this is hardly a 0-day bug). 

I'm not going to be bothered to comment on all your other comments, but maybe recognize that it is a holiday period for the majority of us, and that also slows down reviews. 

You can comment on timing issues or not having enough reviewers look at your stuff, but to retort that gerrit is for show or to let bugs pass through is just insulting.

Patch-set: 1