Update patch set 1
Patch Set 1:

It doesn't seem to matter who one adds as a reviewer. Or how many times a patch is bumped. Nothing rarely happens. I was really looking forward to this being improved with the new company and all, especially now when people are actually getting paid for taking care of this.

Holidays or not, it's been like this for quite a while.

Patch-set: 1