Update patch set 1
Patch Set 1:

And ciwrl, no I'm not trying to insult anyone. I was expressing my view on the situation, albeit maybe a little influenced by my frustration, but it was definitely not my intention to insult anyone.

Maybe it would be a good idea to publish a list of all those that has global +2, so we know who to add as reviewers?

Patch-set: 1