Update patch set 1
Patch Set 1:

I take no issue with constructive criticism, and the scaling issue with regards to review time has been something we have openly acknowledged in the past and is indeed an issue we are looking to solve. But this was the wrong course of action to take. 

Yes, that list could be beneficial to many - I can work on that tomorrow.

Patch-set: 1