Update patch set 1
Patch Set 1: Code-Review-1

Also, if you ever find yourself in this situation again (no stack trace in Android), what I did was run the test while spamming cat /proc/`pgrep xts`/maps, and letting it crash.  Then get the PC address and find it in the memory map to find the library it crashed in. You can then take the base address of the library and subtract it from the PC to get the offset and run arm-eabi-addr2line on it to get the line.

Patch-set: 1