Update patch set 2
Patch Set 2:

So, for lack of a better place to have this discussion, let us start it here and we can move it to a email/forum thread if need be. 

The issue isn't that no one is bothering to review, etc - we see the notifications just the same as anything else. The issue is that first, this patch and all patches for 'legacy' devices by their nature do not benefit actively supported devices. That's the first point. 

Second, even simple patches such as this _do_ affect current support - every time we merge in new upstream code, we now have to account for these legacy patches. 

So that becomes to 'strikes' against these sorts of patches: not only does it not assist in active officially supported CM devices, it hinders/effects future code merges from upstreams. This is why you see hesitance in merging said patches. 

It is my understanding that there is a 'legacy' project with its own gerrit. While I appreciate that you all are taking the time to submit these patches back to CM mainline, I don't think they will (or should) be accepted for the above reasons. It may not be pleasant, but my suggestion is to fork the affected projects and merge in upstream from CM weekly or w/e basis works best (same as we do for CAF code).

Patch-set: 2