Update patch set 2
Patch Set 2:

I'll tell you what: The code is bothering upstream merges, that's why it is removed. I understand your frustration since I own a device needing legacy stuff too, but it's a decision that was made and we cannot change it.

If your device ever qualifies for official builds, I'm sure the changes will be happily merged again.

Patch-set: 2