Update patch set 1
Patch Set 1:

You are fast. I didn't thought that you are merged this change before it gets merged in here.

Dazzozo (the device maintainer) uses your android_frameworks_av to build the CM11 nightlys for U8815.

We have a problem with Viber and other voip apps and thought that maybe this change can solve the problem, but it is not.

Patch-set: 1