Update patch set 4
Patch Set 4:

Hey Dan,

This is the worst customer support ive ever seen! I told you bugs and you tell me to fuck off? not cool.

How do you expect users of the nightlies to contribute to fix the issues if the dev team is unwilling to listen to the problems?

Patch-set: 4