Update patch set 4
Patch Set 4:

Ryan - "Fix the damn dialer" is hardly a bug report. Especially in the comments of a patch that has not only zero to do with your device, but zero to do with audio in general.

Nobody here owes you anything. Coming in and demanding fixes is not the way to get things done.

Patch-set: 4