Update patch set 1
Patch Set 1: Code-Review+1

Without this change, if you save and then play a HD video on hammerhead it fails to play and media server crashes.

This is probably broken on every device that has vendor blobs that support secure buffers via widevine.

Rather than reverting the commit which broke the behaviour, this commit just hacks around the change and therefore should be safe to cherry-pick to stable which would probably be a good idea even though this problem has existed for the last 3 M- releases.

Patch-set: 1