Update patch set 1
Patch Set 1:

Ethan, finally had the time to test what you suggested. I couldn't even get it to compile without defining where the code is to supposed to be placed exactly. How would it find it's way inside namespace android and class CameraParameters otherwise?

IMHO this is good to go as it is, at least it works as advertised.

Patch-set: 1