Update patch set 1
Patch Set 1:

We'll want this to be an ifdef then. Even if QC gets it wrong and sticks a new virtual method in the middle of the class, OEMs will pick it up and that's what they will use, rendering fix something that actually breaks things. For Bacon, we build these blobs ourselves and they match the framework, but this change would break Bacon as is today until we pushed new ones out.

Patch-set: 1