Update patch set 1
Patch Set 1:

Doing this as a board config is horrendous.

With the hammerhead board config, I found 13 Android.mk's in frameworks/av that need to be changed, 1 in frameworks/wilhelm, 2 in external/stagefright-plugins. 1 in frameworks/base and maybe one or two more after that.

(I just added a #error in the ifndef case to blow up each time it was included)

The slightest change in other board config settings could easily cause other includes that I've missed.

I'm going to do this as a global CFLAG and hopefully that will make this workable.

Patch-set: 1