Update patch set 2
Patch Set 2:

Steve,

Is there any reason that this cannot be merged now?

This is needed to fix a bug on hammerhead as well as a new device, should be quite safe and cause little to no future merge conflicts.

I am hoping to get this addressed one way or another before the next M release if that's at all possible.

Patch-set: 2