Update patch set 3
Patch Set 3:

New patch fixes several bugs and also enables the entire path to use 24-bit samples without requiring offload. The upside of this is that it allows mixing of 24-bit streams with any other stream, the downside is that it depends on the buffer reformatting support in AudioFlinger, which may have lurking edge cases.

Still WIP.

Patch-set: 3