Update patch set 1
Patch Set 1:

Yes, it's super odd.  Upon further inspection I found that it was not based on project root, but root/include.  Thus I think the problem is that the project's include path is not being added with -I.  However this is wrong to begin with.  IMHO projects should not have an include folder, just put the include files next to the c files.  There should be a make target that installs the files that should go into the system include folder.  I wonder if there are examples of how other projects handle this?

Patch-set: 1