Update patch set 1
Patch Set 1:

Not sure, I gave up on that and reverted to Google's behavior. because it has never worked correctly. The intent of the original code was to make deep buffer the default mode to save power, but that's mostly irrelevant with a fully working offload implementation.

You need to set AUDIO_FEATURE_LOW_LATENCY_PRIMARY:=true and set up the audio_policy.conf correctly (see bacon) to re-enable low-latency as the primary path (Google behavior).

Patch-set: 1