Update patch set 1
Patch Set 1:

Ethan, if you could check this and the other patches belonging to the LA.AF topic, that would be great. I've ifdef'd some of them for legacy alsa, but let me know if more of them needs it. I think it should be ok as it is though.

Patch-set: 1