Update patch set 1
Patch Set 1:

Nice catch.

The issue that the addition of getCachedDuration is trying to resolve is that after a seek in Play Music (streaming), the duration returned is briefly zero. This causes offload fallback, but worse it causes NuPlayer to choose the low-latency stream instead of the deep-buffer stream due to the check in openAudioSink. Since effects on the global mix (visualizer) don't include low-latency streams, the visualizer (and potentially other effects) would no longer function after a seek.

The whole code flow here is nasty and error-driven.

The CTS case which was failing previously was "testResumeAtEnd" in MediaPlayerTest, which does a seek to the middle of the stream before playing, and sets up a callback that triggers at EOS to restart playback from the beginning. The initial seek failed so the EOS never came before the timer expired :)

I'll give it a run with this.

Patch-set: 1