Update patch set 2
Patch Set 2:

S4 Mini (msm8930) with both legacy and new audio HAL, and my old Galaxy W (msm7x30). It's easy to reproduce testing keyboard with touch tones enabled. If you're not seeing it on B family devices, it might be because it depends on some kernel side stuff that the A family doesn't have.

Patch-set: 2