Update patch set 9
Patch Set 9:

I don't think INCALL_VOICE should be the guard for that, mostly because I don't think there _has_ to be a hard dependency on it. If we're using it as a proxy for ifeq B_FAMILY we might as well be explicit about it.

That said, this is fine.

Patch-set: 9