Update patch set 1
Patch Set 1:

Okay, take a look at this commit: http://review.cyanogenmod.org/#/c/27703/, Andrew changed the FIMC1 allocation to the ION region to match the behavior of our blobs. This way, the hardware codecs run as they should, but if we try to run FIMC here, the HWC cannot buffer the memory properly because it's not in the FIMC region, that's why it fails very hard at the runFimc funtion.
In that same way, we cannot try to run ION here because our device uses FIMC to process memory to memory allocations

Patch-set: 1