Update patch set 1
Patch Set 1:

Ethan (or anyone reviewing the code), I commented the #ifdef lines for a few things that were already there to increase code portability.
Some of the changes needed by this were already part of some CAF commit, but I'd like this commit to be applicable on top of an AOSP branch, thus the hints about needed code with commented ifdefs.

Patch-set: 1