Update patch set 2
Patch Set 2:

I thought so as well on first sight, but then again it's only the header files that are open source here, the real extractor code is shipped in form of statics (*.a), so I guess it's okay.
Also, if Hardkernel is allowed to make it public, so should we.

Patch-set: 2