Update patch set 2
Patch Set 2:

I don't want to be a party pooper, but weather it's the header files or that .a's, it seems that we don't have the license to distribute these. Being "just" header files or binaries doesn't mean that they don't have copyright. I can't speak about the *a's, but, until we see something saying otherwise, we'll should assume that the notice on the header file is correct, and we, indeed, have no permission to distribute it.

We shouldn't assume HardKernel has license to distribute this just because they are doing it, and if they have it, it doesn't mean that we do. And taking into account Ittiam's line of business, I wouldn't be surprised to know we really can't.

td;dr;: IMHO, to the best of our knowledge, we can't ship this. Trying to figure this out next to Hardkernel or ITTIAM seems a possible line of action.

Patch-set: 2