Update patch set 11
Patch Set 11:

(1 comment)

I agree with Marcos. We're adding unnecessary changes for the sake of maintaining some portability. This isn't really desired since we already have to worry about the maintaining compatibility across multiple architectures and platforms in this code. 

Also, are you using blobs for the AAC/MP3 decoders? We currently don't have the source code for this, it's dependent on a static libsrp of some sort.

Patch-set: 11