Update patch set 11
Patch Set 11:

Ethan, I'm building all my OMX libraries from source. I'm not using the BSP that CM currently has, but the newer one (4.4 from Hardkernel) that I attempted to bring in a while back, but you were saying the code layout changed too much and honestly, I was too lazy to reformat all the code so you would accept it.
The new BSP is working fine though, if you want, you can fork it over to CM from here:
https://github.com/TeamEOS?utf8=?&query=slsi.
I'm currently working on Exynos5420 and 5433, but adding support for your 5410 devices should be easy.

Regarding this patch, would it be any better if I removed the ifdefs and instead commented the respective lines? I really don't want this to be merged without any trace on the code that was introduced by earlier patches that this one depends on.

Patch-set: 11