Update patch set 6
Patch Set 6:

I honestly only took your patchset and improved it, because your style of moving code was invasive and rather undesirable.
If you have major criticism on the code itself, that is your area to improve, as I only worked with the code you uploaded.
Also, for your current method of checking for rear/front camera, being a member function or not shouldn't matter as long as the correct parameters are passed.

Feel free to revert to an earlier patchset and offer your solution again, I'm curious what the CM guys will say.

Patch-set: 6