Update patch set 1
Patch Set 1:

The previous commit should fix the root issue but I thought that this was a nice improvement to make sure that we can recover when we hit a bad case.

If we merge the previous commit but opt to not merge this one, we should update the second hard coded copy of 256*1024 which should have been updated to 1024*1024 when kGrayArea was updated to work around buffering problems.

Patch-set: 1