Update patch set 1
Patch Set 1:

The specific anime video plays fine on victara with offloading enabled (it may just be that my internet connection is too fast right now to reproduce it or the audio/video decoders are staying slightly closer together and making it faster to get the missing data loaded).

In the logs though I do see:

08-25 04:56:36.919 I/NuCachedSource2(10236): Cache full, done prefetching for now
08-25 04:56:53.474 I/NuCachedSource2(10236): restarting prefetcher, totalSize = 5111808
08-25 04:56:53.534 I/NuCachedSource2(10236): new range: offset= 29108054

which indicates that when the prefetcher is started, we drop all but the previous 1MB of data and then immediately do a seek which causes us to discard all the prefetched data.

Patch-set: 1