Update patch set 2
Patch Set 2:

> Certainly interesting. But I'm not sure how useful it will be to
 > have support in the framework for those functions. Because we'd
 > need camera apps that make use of those functions. And AFAIK only
 > the camera apps from the various OEMs do. Or am I missing
 > something?
 > 
 > I implemented support for set_custom_parameter in the
 > htc/tegra3-common camera wrapper. Unfortunately I still couldn't
 > get HDR and burst to work well.
 > Also did you forget to upload the change to hardware/libhardware?

I can push it, but it's the same as what you have in tegra3-common. This was more of a reference as I'm not sure how safe it is to actually push this code publicly. HP made a mistake and uploaded their whole android 4.2.2 build source and it contains every single change from Nvidia. If the wrapper is a better idea I will go ahead and attempt to use it as the camera is completely broken on the HP late 21. If you want a link to the zip, let me know as it also includes a lot of nvidia HAL you may want to take a look at including camera.

Patch-set: 2
Reviewer: Carl Draper <5901@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0