Update patch set 3
Patch Set 3: Code-Review-1

I'd appreciate some input on this. The check I do is quite insane and not proper but I feel that something along this line could get incorporated into the already available checks.

Also many devices had offload issues, sadly I do not have any logs or input on what 'they' are experiencing and if it is the same issue.

Patch-set: 3