Update patch set 4
Patch Set 4:

FYI, Samsung released 5.1 on another tab pro which likely shares the same exynos5420 kernel with my picassowifi.  I tried updating the gpu and video drivers to the latest version and this didn't fix the problem.

At this point I'm guess a blob issue and nothing that I can fix in the kernel.  Hopefully we'll get the 5.1 release as well as and that will magically fix it for me...

Patch-set: 4