Update patch set 4
Patch Set 4: Code-Review-2

So the "out of space" thing is supposed to happen.

1. Android reads track from storage, pumps to DSP
2. DSP says "HEY STOP IT NO MORE MEMORY" and logs the out of space message
3. Android goes to sleep
4. DSP gets close to finishing playout of the buffer, wakes up Android and asks for more data
5. Repeat until done.

The reason AudioFX isn't be applied is because in offload mode, the DSP decompresses the file and plays it. Therefore, the DSP must also apply effects.  It's not working because your DSP doesn't support it (probably due to a missing topology or because it's just old). By invalidating the track, you kill offload and ask Android to do the effects, which run on the CPU.

Patch-set: 4
Reviewer: Steve Kondik <32@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=-2