Update patch set 1
Patch Set 1:

I worry about people blindly using this.  "I need to add 1 symbol, let's copy all of this into my device and add the one symbol" even though there is already a better way to handle that.

However, the way HTC arranged those two classes the options are pretty limited.  The only other thing I could think of would be to add a

#define CAMERA_PARAMETERS_SUPERCLASS CameraParameter_EXT

to the existing #include "camera/CameraParametersExt.h"

However, I'm not sure if that is sufficient for hima because I'm too lazy to see if there are any other diffs in the existing CameraParameters code that would make it impossible to keep it isolated entirely to the superclass.

Patch-set: 1
Label: Code-Review=0