Update patch set 1
Patch Set 1:

It's kind of weird, most of the Insignal repos do not have any specific edits for the 4.4.4 branches, so I guess he either forgot to push them or they are simply not available right now. The 4.4.2 branches are what I'm looking at mostly, those seem to be the latest sauce.

Patch-set: 1