Update patch set 1
Patch Set 1:

Hmm, does audio.a2dp even work correctly on A family, since it doesn't support separate output? If it's desirable to use it, it's probably better migrating to B family HAL.

I'll change the ifdef in AudioOutputDescriptor, so it's more like what it was is previous CM, minus the logic since I couldn't port that from AudioPolicyManager.

Patch-set: 1