Update patch set 1
Patch Set 1:

The issue was introduced in r26 merge, with the following patch: https://github.com/CyanogenMod/android_frameworks_av/commit/a2ab4505c807f42afe34809409c4e85d91618a8c

Never was an issue in any CM version, before this. As the comment in the patch says, it's a "workaround for first HAL write being CPU bound on some devices". Which I interpret as "Yeah it might fix it on some devices, but it's irrelevant on others". Google doesn't care though, as long as it works on the Nexuses.

I think part of the problem that causes the pop noise, is that the A family HAL only properly supports PCM16 format, while I believe the mixer thread uses PCM_FLOAT.

I'll see if playing with the latency values makes any difference.

Patch-set: 1