Update patch set 2
Patch Set 2:

Well, it's the old ass A family HAL. :) It's a wonder it's even still working. If someone is able to and willing, feel free to try getting its outdated policy manager working again. :)

Personally, I could just switch to the B family HAL and be done with it. I've already done all the work migrating from snd_soc config to the new xml ones. But realistically, I don't think we can expect everyone to do that.

Patch-set: 2