Update patch set 1
Patch Set 1:

Steve, there's no input from the microphone when recording from for example the SoundRecorder app. Just total silence. I think someone mentioned that Google voice search is also broken.

This is with the A family HAL. I've tested the B family HAL on the same device, and no problems there. A long time ago when Google switched to PCM_FLOAT for output, CAF made a change in f/av for A family to use PCM_16_BIT, and stated that that's the only format the HAL supported. I think there's something similar here now with input.

Patch-set: 1