Update patch set 4
Patch Set 4:

I'd much rather make the decoder use the standard OMX color formats, but I'm not sure if that's possible.
Also, the handling of OMX_SEC_COLOR_FormatYVU420Planar is rather clear, but not so much for OMX_SEC_COLOR_FormatNV21Linear.
There seems to be little documentation on the NV21 linear format that Samsung wants to use here.

Patch-set: 4