Update patch set 2
Patch Set 2:

I mean the commit message is quite clear, or do I miss something here? get_presentation_position was not supported for pre Kitkat audio blobs and therefore assumed to be NULL. But on endeavoru with the Tegra3 blobs it was not NULL and causing crashes when called. I never investigated where it was pointing. So the fix was to not override getTimestamp in AudioStreamOutSink and therefore returning INVALID_OPERATION every time.
If I recall it correctly that wasn't needed anymore when we figured out how to use the Kitkat blobs.
I guess it's not needed anymore.

Patch-set: 2