Update patch set 3
Patch Set 3:

I think we probably want to return the remote process UID and also protect the callback registration with a permission. It's not like the session ids are secure or anything, but it's a notable omission in stock Android for whatever reason (apps have to send a control intent).

Patch-set: 3