Update patch set 4
Patch Set 4:

FYI: To use this stuff, you need to add an output_session_processing block to your audio_effects.conf which defines the list of effects to attach to a stream type (with no configuration). Will do an OSS example once I'm satisfied with it on Bacon with Waves.

Patch-set: 4