Update patch set 3
Patch Set 3:

No, this applies specifically to stuff declared in the conf file under output_session_processing. What this does is auto-attach effects. A previous patch from me added a callback mechanism to let apps like AudioFX get notified when this happens. The purpose of all this is to attach effects to streams automatically, even if the application doesn't send a control intent (which is a lot of them).

MaxxAudio does some pretty heavyweight setup since it is handling what is effectively calibration data for output devices, and has to send hundreds of parameters to the backend. This patch eliminates the latency from this, and constrains the attach to streams with specific attributes.

Patch-set: 3