Update patch set 1
Patch Set 1:

Such plumbing, much binder!

This is a refcounted version of the previous work, which triggers at stream open and stream close. We just need to save off a few things when getOutputForAttr() is called. This fixes the issues with sessions either being removed immediately or not at all.

Additionally, I found myself writing some completely horrible code in ModioFX to persist the session list, but it's essentially an impossible task in light of the LMK. The symptom of this is that you open the app and none of the controls will actually do anything, which is not good. So this patch adds a query interface, which will have a Java counterpart shortly. This allows you to get a list of all active audiosessions including the metadata.

Java part coming next.

Patch-set: 1