Update patch set 1
Patch Set 1:

It's my own port for the Elephone Trunk.

There weren't any crashes until commit 80c362f, but afterwards mediaserver crashes after recording a video. I didn't find anything in this commit that seems to be related to that problem, so it might as well be that this pointer gets overwritten accidentally by some other, unrelated, array out of bounds operation.

Patch-set: 1