Don't try to connect camera service if it is disabled.
On Android Wear devices we don't have cameras, so we don't need the
proxy service as well. If it is disabled by the system property, don't
wait for it to start, but rather return null pointer in
getCameraService(), which causes getNumberOfCameras() to report zero
cameras available.
The same logic applies to ACameraManager, where we return an empty list.

Bug: 28560707
Change-Id: I4c0bc29f061f1b66710c8188a7916bfaf089d23f