stagefright: unblock stop if already in error state
If we received an error from component during stopping, or we
already have a sticky error when we received a stop request,
post reply with error immediatly to unblock the client.

Internally, we process the stop as usual and stay in STOPPING
state. If the client doesn't do anything after that, we wait
for the component's shutdown complete. (It's possible it never
completes, and we leak the resouce; but that's still better than
leaking the resource AND have the app ANR.) If the app initiates
release or reset, we'll release with timeout.

bug: 31671152
Change-Id: I8b7ada773c1fff77aff736efd40087bd3e9b178e