Update patch set 8
Patch Set 8:

Eric and Alberto this should be the last step for surnia vid recording along with TARGET_NEEDS_PLATFORM_TEXT_RELOCATIONS := true in boardconfig so libffmpeg etc will load and not spaz out.

The logcats were ridiculous because they don't like to be told no; I was hoping the last piece to puzzle was in tha framework.

Thanks Simon.

Patch-set: 8