Update patch set 11
Patch Set 11:

I think the security implications should be outlined a little clearer in the commit message. For devices that have at least some access to their camera HAL and codec code this is generally fixable without a major overhaul.

CAF has examples, and here is one for OMAP: http://gerrit.unlegacy-android.cf/#/c/395/

The framework serializes the metadata pointers and fds for you if you're using the right structure for it.

Patch-set: 11